+++ src/lib/linear_model/linear_regression.ts
@@ -4,12 +4,26 @@
  */
 import * as tf from '@tensorflow/tfjs';
 import { size } from 'lodash';
-import { validateMatrix1D } from '../ops';
-import { Type1DMatrix } from '../types';
+import * as numeric from 'numeric';
+import { inferShape, reshape } from '../ops';
+import { Type1DMatrix, Type2DMatrix } from '../types';
 import math from '../utils/MathExtra';
 
 /**
+ * Type of Linear Regression
+ * Univariate = It can handle a 1 dimensional input
+ * Multivariate = It can handle a 2 dimensional input
+ * @ignore
+ */
+export enum TypeLinearReg {
+  UNIVARIATE = 'UNIVARIATE',
+  MULTIVARIATE = 'MULTIVARIATE'
+}
+
+/**
  * Ordinary least squares Linear Regression.
+ *
+ * It supports both univariate and multivariate linear regressions.
  *
  * @example
  * import { LinearRegression } from './linear_regression';
@@ -17,12 +31,19 @@ import math from '../utils/MathExtra';
  * const X = [1, 2, 4, 3, 5];
  * const y = [1, 3, 3, 2, 5];
  * linearRegression.fit(X, y);
- * console.log(lr.predict([1, 2]));
+ * lr.predict([1, 2]);
  * // [ 1.1999999999999995, 1.9999999999999996 ]
+ *
+ * const linearRegression2 = new LinearRegression();
+ * const X2 = [[1, 1], [1, 2], [2, 2], [2, 3]];
+ * const y2 = [1, 1, 2, 2];
+ * linearRegression2.fit(X2, y2);
+ * lr.predict([[1, 2]]);
+ * // [1.0000001788139343]
  */
 export class LinearRegression {
-  private b0: number;
-  private b1: number;
+  private weights: number[] = [];
+  private type: TypeLinearReg = TypeLinearReg.MULTIVARIATE;
 
   /**
    * fit linear model
@@ -30,78 +51,158 @@ export class LinearRegression {
    * @param {any} y - target values
    */
   public fit(
-    X: Type1DMatrix<number> = null,
-    y: Type1DMatrix<number> = null
+    X: Type1DMatrix<number> | Type2DMatrix<number> = null,
+    y: Type1DMatrix<number> | Type2DMatrix<number> = null
   ): void {
-    const xShape = validateMatrix1D(X);
-    const yShape = validateMatrix1D(y);
+    const xShape = inferShape(X);
+    const yShape = inferShape(y);
     if (xShape.length === 1 && yShape.length === 1 && xShape[0] === yShape[0]) {
-      this.coefficients(X, y); // getting b0 and b1
+      // Univariate linear regression
+      this.type = TypeLinearReg.UNIVARIATE;
+      this.weights = this.calculateUnivariateCoeff(X, y); // getting b0 and b1
+    } else if (
+      xShape.length === 2 &&
+      yShape.length === 1 &&
+      xShape[0] === yShape[0]
+    ) {
+      this.type = TypeLinearReg.MULTIVARIATE;
+      this.weights = this.calculateMultiVariateCoeff(X, y);
     } else {
       throw new Error(
         `Sample(${xShape[0]}) and target(${yShape[0]}) sizes do not match`
       );
     }
   }
-
   /**
    * Predict using the linear model
    * @param {number} X - Values to predict.
    * @returns {number}
    */
-  public predict(X: Type1DMatrix<number> = null): number[] {
-    validateMatrix1D(X);
-    const preds = [];
-    for (let i = 0; i < size(X); i++) {
-      preds.push(this.b0 + this.b1 * X[i]);
+  public predict(
+    X: Type1DMatrix<number> | Type2DMatrix<number> = null
+  ): number[] {
+    const xShape = inferShape(X);
+    if (
+      xShape.length === 1 &&
+      this.type.toString() === TypeLinearReg.UNIVARIATE.toString()
+    ) {
+      return this.univariatePredict(X as Type1DMatrix<number>);
+    } else if (
+      xShape.length === 2 &&
+      this.type.toString() === TypeLinearReg.MULTIVARIATE.toString()
+    ) {
+      return this.multivariatePredict(X as Type2DMatrix<number>);
+    } else {
+      throw new TypeError(
+        `The matrix is incorrectly shaped: while X is ${
+          xShape.length
+        }, type is ${this.type.toString().toLowerCase()}`
+      );
     }
-    return preds;
   }
-
   /**
    * Get the model details in JSON format
    */
   public toJSON(): {
     /**
-     * Coefficients component: b0
+     * Coefficients
      */
-    b0: number;
+    weights: number[];
     /**
-     * Coefficients component: b1
+     * Type of the linear regression model
      */
-    b1: number;
+    type: TypeLinearReg;
   } {
     return {
-      b0: this.b0,
-      b1: this.b1
+      weights: this.weights,
+      type: this.type
     };
   }
 
   /**
    * Restore the model from a checkpoint
-   * @param {number} b0 - Coefficients component: b0
-   * @param {number} b1 - Coefficients component: b1
    */
-  public fromJSON({ b0 = null, b1 = null }: { b0: number; b1: number }): void {
-    if (!b0 || !b1) {
+  public fromJSON({
+    /**
+     * Model's weights
+     */
+    weights = null,
+    /**
+     * Type of linear regression, it can be either UNIVARIATE or MULTIVARIATE
+     */
+    type = null
+  }: {
+    weights: number[];
+    type: TypeLinearReg;
+  }): void {
+    if (!weights || !type) {
       throw new Error(
-        'You must provide both b0 and b1 to restore Linear Regression'
+        'You must provide both weights and type to restore the linear regression model'
       );
     }
-    this.b0 = b0;
-    this.b1 = b1;
+    this.weights = weights;
+    this.type = type;
   }
 
   /**
-   * Calculates coefficient for linear regression
+   * Univariate prediction
+   * y = b0 + b1 * X
+   *
+   * @param X
+   */
+  private univariatePredict(X: Type1DMatrix<number> = null): number[] {
+    const preds = [];
+    for (let i = 0; i < size(X); i++) {
+      preds.push(this.weights[0] + this.weights[1] * X[i]);
+    }
+    return preds;
+  }
+
+  /**
+   * Multivariate prediction
+   * y = (b0 * X0) + (b1 * X1) + (b2 * X2) + ....
+   *
+   * @param X
+   */
+  private multivariatePredict(X: Type2DMatrix<number> = null): number[] {
+    const preds = [];
+    for (let i = 0; i < X.length; i++) {
+      const row = X[i];
+      let yPred = 0;
+      for (let j = 0; j < row.length; j++) {
+        yPred += this.weights[j] * row[j];
+      }
+      preds.push(yPred);
+    }
+    return preds;
+  }
+
+  /**
+   * Calculates univariate coefficients for linear regression
    * @param X - X values
    * @param y - y targets
    */
-  private coefficients(X, y): void {
-    // TODO: Fix the typing
+  private calculateUnivariateCoeff(X, y): number[] {
     const xMean: any = tf.mean(X).dataSync();
     const yMean: any = tf.mean(y).dataSync();
-    this.b1 = math.covariance(X, xMean, y, yMean) / math.variance(X, xMean);
-    this.b0 = yMean - this.b1 * xMean;
+    const b1 = math.covariance(X, xMean, y, yMean) / math.variance(X, xMean);
+    const b0 = yMean - b1 * xMean;
+    return this.weights.concat([b0, b1]);
+  }
+
+  /**
+   * Calculate multivariate coefficients for linear regression
+   * @param X
+   * @param y
+   */
+  private calculateMultiVariateCoeff(X, y): number[] {
+    const [q, r] = tf.linalg.qr(tf.tensor2d(X));
+    const rawR = reshape(Array.from(r.dataSync()), r.shape);
+    const weights = tf
+      .tensor(numeric.inv(rawR))
+      .dot(q.transpose())
+      .dot(tf.tensor(y))
+      .dataSync();
+    return Array.from(weights);
   }
 }